// src/app/modulos/logistica-vehicular/detalle/evaluacion/evaluacion.enum.ts
export const AGREGAR = 'AGREGAR';

export const EDITAR = 'EDITAR';

export const RESPALDO = 'RESPALDO';

export enum ModalTipoVerificacion {
  AGREGAR = 'AGREGAR',
  EDITAR = 'EDITAR'
}

export enum DocumentosBotonesServicio {
  ENVIO_COMITE = 'enviarComite',
  APROBADO = 'aprobarEvaluacion',
  OBSERVADO = 'observarEvaluacion',
  SIGUIENTE = 'siguienteBloque',
  ANULAR = 'anularEvaluacion'
}